Simplify RadioTest and made radios required
It was unclear what was being tested, and yet there was a load of duplication from InputTest.
Radios should be required in the same way checkboxes should be optional. One thing to think about
in making radios required is how it affects the validity of groups that are otherwise blank.